+++ app/ui/home/map/index.tsx
@@ -3,11 +3,11 @@ import dynamic from "next/dynamic";
 
 
 const Map = dynamic(
-    () => import('@/app/ui/home/map'),
+    () => import('@/app/ui/home/map/map'),
     { 
       loading: () => <p>A map is loading</p>,
       ssr: false
     }
 )
 
-  export default Map;
+export default Map;