+++ app/home/page.tsx
@@ -9,8 +9,12 @@ export default async function Page() {
 
     return (
         <>
-        <Map/>
-        <RangeGraph characters={lighthouses}/>
+         <div className="grid grid-cols-3 grid-rows-3 gap-4">
+            <div className="col-span-2 row-span-4"><Map/></div>
+            <div className="col-span-1 row-span-1"><RangeGraph characters={lighthouses}/></div>
+            <div className="col-span-1 row-span-1"><RangeGraph characters={lighthouses}/></div>
+            <div className="col-span-1 row-span-1"><RangeGraph characters={lighthouses}/></div>
+         </div>
         </>
     )
 }