+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## verson 0.2.0
 Added linegraph to Historical page to show wind metrics
 Added linegraph to Marine page to show waveheight metrics
+Added map to Forecast page
+Changed lighthouse data from database fetch to placeholder-data for map on home page
+Added Chart to Home page
 
 ## verson 0.1.1
 Added map to home page